Make Danger run
